@@ -35,10 +35,10 @@ class RangeDecoder {
   static const int _kBitModelTotal = 1 << _kNumBitModelTotalBits;
   static const int _kNumMoveBits = 5;
 
-  int _range;
-  int _code;
+  late int _range;
+  late int _code;
 
-  InStream _stream;
+  InStream? _stream;
 
   void setStream(InStream stream) {
     _stream = stream;
@@ -52,22 +52,22 @@ class RangeDecoder {
     _code = 0;
     _range = -1;
 
-    for (var i = 0; i < 5; ++ i) {
-      _code = (_code << 8) | _stream.read();
+    for (var i = 0; i < 5; ++i) {
+      _code = (_code << 8) | _stream!.read();
     }
   }
 
   int decodeDirectBits(int numTotalBits) {
     var result = 0;
 
-    for (var i = numTotalBits; i > 0; -- i) {
+    for (var i = numTotalBits; i > 0; --i) {
       _range = (_range >> 1) & 0x7fffffff;
       var t = ((_code - _range) >> 31) & 1;
       _code -= _range & (t - 1);
       result = (result << 1) | (1 - t);
 
       if ((_range & _kTopMask) == 0) {
-        _code = (_code << 8) | _stream.read();
+        _code = (_code << 8) | _stream!.read();
         _range <<= 8;
       }
     }
@@ -75,17 +75,17 @@ class RangeDecoder {
     return result;
   }
 
-  int decodeBit(List<int> probs, int index) {
-    int prob = probs[index];
+  int decodeBit(List<int?> probs, int index) {
+    int prob = probs[index]!;
 
-    var newBound = ((_range >>_kNumBitModelTotalBits) & 0x1fffff) * prob;
+    var newBound = ((_range >> _kNumBitModelTotalBits) & 0x1fffff) * prob;
 
-    if ((new Int32(_code) ^ 0x80000000) < (new Int32(newBound) ^ 0x80000000)) {
+    if ((Int32(_code) ^ 0x80000000) < (Int32(newBound) ^ 0x80000000)) {
       _range = newBound;
       probs[index] = prob + ((_kBitModelTotal - prob) >> _kNumMoveBits);
 
       if ((_range & _kTopMask) == 0) {
-        _code = (_code << 8) | _stream.read();
+        _code = (_code << 8) | _stream!.read();
         _range <<= 8;
       }
 
@@ -97,16 +97,16 @@ class RangeDecoder {
     probs[index] = prob - ((prob >> _kNumMoveBits) & 0x7ffffff);
 
     if ((_range & _kTopMask) == 0) {
-      _code = (_code << 8) | _stream.read();
+      _code = (_code << 8) | _stream!.read();
       _range <<= 8;
     }
 
     return 1;
   }
 
-  static void initBitModels(List<int> probs) {
-    for (var i = 0; i < probs.length; ++ i) {
-      probs[i] = _kBitModelTotal >> 1;
+  static void initBitModels(List<int> probs, int length) {
+    for (var i = 0; i < length; ++i) {
+      probs.replaceOrAdd(_kBitModelTotal >> 1, i);
     }
   }
 }
@@ -117,14 +117,14 @@ class RangeEncoder {
   static const int _kBitModelTotal = 1 << _kNumBitModelTotalBits;
   static const int _kNumMoveBits = 5;
 
-  OutStream _stream;
+  OutStream? _stream;
 
-  Int64 _low;
-  int _range;
-  int _cacheSize;
-  int _cache;
+  late Int64 _low;
+  late int _range;
+  late int _cacheSize;
+  late int _cache;
 
-  int _position;
+  late int _position;
 
   void setStream(OutStream stream) {
     _stream = stream;
@@ -143,37 +143,37 @@ class RangeEncoder {
   }
 
   void flushData() {
-    for (var i = 0; i < 5; ++ i) {
+    for (var i = 0; i < 5; ++i) {
       shiftLow();
     }
   }
 
   void flushStream() {
-    _stream.flush();
+    _stream!.flush();
   }
 
-  void shiftLow(){
+  void shiftLow() {
     var lowHi = _low.shiftRightUnsigned(32).toInt();
     if ((lowHi != 0) || (_low < 0xff000000)) {
       _position += _cacheSize;
       var temp = _cache;
       do {
-        _stream.write((temp + lowHi) & 0xff);
+        _stream!.write((temp + lowHi) & 0xff);
         temp = 0xff;
-      } while (-- _cacheSize != 0);
+      } while (--_cacheSize != 0);
       _cache = _low.toInt32().shiftRightUnsigned(24).toInt();
     }
-    ++ _cacheSize;
+    ++_cacheSize;
     _low = (_low & 0xffffff) << 8;
   }
 
   void encodeDirectBits(int v, int numTotalBits) {
-    for (var i = numTotalBits - 1; i >= 0; -- i) {
+    for (var i = numTotalBits - 1; i >= 0; --i) {
       _range = (_range >> 1) & 0x7fffffff;
       if (((v >> i) & 1) == 1) {
         _low += _range;
       }
-      if ((_range & _kTopMask) == 0){
+      if ((_range & _kTopMask) == 0) {
         _range <<= 8;
         shiftLow();
       }
@@ -185,14 +185,14 @@ class RangeEncoder {
   static const int _kNumMoveReducingBits = 2;
   static const int _kNumBitPriceShiftBits = 6;
 
-  static void initBitModels(List<int> probs) {
-    for (var i = 0; i < probs.length; ++ i) {
+  static void initBitModels(List<int?> probs) {
+    for (var i = 0; i < probs.length; ++i) {
       probs[i] = _kBitModelTotal >> 1;
     }
   }
 
-  void encode(List<int> probs, int index, int symbol) {
-    int prob = probs[index];
+  void encode(List<int?> probs, int index, int symbol) {
+    int prob = probs[index]!;
 
     var newBound = ((_range >> _kNumBitModelTotalBits) & 0x1fffff) * prob;
 
@@ -200,7 +200,7 @@ class RangeEncoder {
       _range = newBound;
       probs[index] = prob + ((_kBitModelTotal - prob) >> _kNumMoveBits);
     } else {
-      _low += new Int64(0xffffffff) & newBound;
+      _low += Int64(0xffffffff) & newBound;
       _range -= newBound;
       probs[index] = prob - (prob >> _kNumMoveBits);
     }
@@ -210,52 +210,54 @@ class RangeEncoder {
     }
   }
 
-  static final List<int> _probPrices = _buildProbPrices();
+  static final List<int?> _probPrices = _buildProbPrices();
 
   static List<int> _buildProbPrices() {
-    var probPrices = new List<int>(_kBitModelTotal >> _kNumMoveReducingBits);
+    var probPrices = <int>[];
 
-    probPrices[0] = 0;
+    probPrices.replaceOrAdd(0, 0);
 
     var kNumBits = _kNumBitModelTotalBits - _kNumMoveReducingBits;
-    for (var i = kNumBits - 1; i >= 0; -- i) {
+    for (var i = kNumBits - 1; i >= 0; --i) {
       var start = 1 << (kNumBits - i - 1);
       var end = 1 << (kNumBits - i);
-      for (var j = start; j < end; ++ j) {
-        probPrices[j] = (i << _kNumBitPriceShiftBits) +
-            (((end - j) << _kNumBitPriceShiftBits) >> (kNumBits - i - 1));
+      for (var j = start; j < end; ++j) {
+        probPrices.replaceOrAdd(
+            (i << _kNumBitPriceShiftBits) +
+                (((end - j) << _kNumBitPriceShiftBits) >> (kNumBits - i - 1)),
+            j);
+        // probPrices[j] = (i << _kNumBitPriceShiftBits) +
+        //     (((end - j) << _kNumBitPriceShiftBits) >> (kNumBits - i - 1));
       }
     }
 
     return probPrices;
   }
 
-  static int getPrice(int prob, int symbol) =>
-    _probPrices[((new Int32(prob - symbol) ^ new Int32(-symbol)).toInt()
-        & (_kBitModelTotal - 1)) >> _kNumMoveReducingBits];
+  static int? getPrice(int prob, int symbol) =>
+      _probPrices[((Int32(prob - symbol) ^ Int32(-symbol)).toInt() &
+              (_kBitModelTotal - 1)) >>
+          _kNumMoveReducingBits];
 
-  static int getPrice0(int prob) =>
-    _probPrices[prob >> _kNumMoveReducingBits];
+  static int getPrice0(int prob) => _probPrices[prob >> _kNumMoveReducingBits]!;
 
   static int getPrice1(int prob) =>
-    _probPrices[(_kBitModelTotal - prob) >> _kNumMoveReducingBits];
+      _probPrices[(_kBitModelTotal - prob) >> _kNumMoveReducingBits]!;
 }
 
 class BitTreeDecoder {
-  final List<int> _models;
+  final List<int> _models = [];
   final int _numBitLevels;
 
-  BitTreeDecoder(int numBitLevels)
-    : _numBitLevels = numBitLevels,
-      _models = new List<int>(1 << numBitLevels);
+  BitTreeDecoder(int numBitLevels) : _numBitLevels = numBitLevels;
 
   void init() {
-    RangeDecoder.initBitModels(_models);
+    RangeDecoder.initBitModels(_models, 1 << _numBitLevels);
   }
 
   int decode(RangeDecoder rangeDecoder) {
     var m = 1;
-    for (var i = _numBitLevels; i > 0; -- i) {
+    for (var i = _numBitLevels; i > 0; --i) {
       m = (m << 1) | rangeDecoder.decodeBit(_models, m);
     }
     return m - (1 << _numBitLevels);
@@ -263,7 +265,7 @@ class BitTreeDecoder {
 
   int reverseDecode(RangeDecoder rangeDecoder) {
     var m = 1, symbol = 0;
-    for (var i = 0; i < _numBitLevels; ++ i) {
+    for (var i = 0; i < _numBitLevels; ++i) {
       var bit = rangeDecoder.decodeBit(_models, m);
       m = (m << 1) | bit;
       symbol |= bit << i;
@@ -271,10 +273,10 @@ class BitTreeDecoder {
     return symbol;
   }
 
-  static int reverseDecode2(List<int>models, int startIndex,
-                            RangeDecoder rangeDecoder, int numBitLevels) {
+  static int reverseDecode2(List<int?> models, int startIndex,
+      RangeDecoder rangeDecoder, int numBitLevels) {
     var m = 1, symbol = 0;
-    for (var i = 0; i < numBitLevels; ++ i) {
+    for (var i = 0; i < numBitLevels; ++i) {
       var bit = rangeDecoder.decodeBit(models, startIndex + m);
       m = (m << 1) | bit;
       symbol |= bit << i;
@@ -284,21 +286,19 @@ class BitTreeDecoder {
 }
 
 class BitTreeEncoder {
-  final List<int> _models;
+  final List<int> _models = [];
   final int _numBitLevels;
 
-  BitTreeEncoder(int numBitLevels)
-    : _numBitLevels = numBitLevels,
-      _models = new List(1 << numBitLevels);
+  BitTreeEncoder(int numBitLevels) : _numBitLevels = numBitLevels;
 
   void init() {
-    RangeDecoder.initBitModels(_models);
+    RangeDecoder.initBitModels(_models, 1 << _numBitLevels);
   }
 
   void encode(RangeEncoder rangeEncoder, int symbol) {
     var m = 1;
     for (var bitIndex = _numBitLevels; bitIndex > 0;) {
-      -- bitIndex;
+      --bitIndex;
       var bit = (symbol >> bitIndex) & 1;
       rangeEncoder.encode(_models, m, bit);
       m = (m << 1) | bit;
@@ -307,7 +307,7 @@ class BitTreeEncoder {
 
   void reverseEncode(RangeEncoder rangeEncoder, int symbol) {
     var m = 1;
-    for (var i = 0; i < _numBitLevels; ++ i) {
+    for (var i = 0; i < _numBitLevels; ++i) {
       var bit = symbol & 1;
       rangeEncoder.encode(_models, m, bit);
       m = (m << 1) | bit;
@@ -319,9 +319,9 @@ class BitTreeEncoder {
     var price = 0;
     var m = 1;
     for (var bitIndex = _numBitLevels; bitIndex > 0;) {
-      -- bitIndex;
+      --bitIndex;
       var bit = (symbol >> bitIndex) & 1;
-      price += RangeEncoder.getPrice(_models[m], bit);
+      price += RangeEncoder.getPrice(_models[m], bit)!;
       m = (m << 1) | bit;
     }
     return price;
@@ -330,32 +330,32 @@ class BitTreeEncoder {
   int reverseGetPrice(int symbol) {
     var price = 0;
     var m = 1;
-    for (var i = _numBitLevels; i > 0; -- i) {
+    for (var i = _numBitLevels; i > 0; --i) {
       var bit = symbol & 1;
       symbol >>= 1;
-      price += RangeEncoder.getPrice(_models[m], bit);
+      price += RangeEncoder.getPrice(_models[m], bit)!;
       m = (m << 1) | bit;
     }
     return price;
   }
 
-  static int reverseGetPrice2(List<int> models, int startIndex,
-      int numBitLevels, int symbol) {
+  static int reverseGetPrice2(
+      List<int?> models, int startIndex, int numBitLevels, int symbol) {
     var price = 0;
     var m = 1;
-    for (var i = numBitLevels; i > 0; -- i) {
+    for (var i = numBitLevels; i > 0; --i) {
       var bit = symbol & 1;
       symbol >>= 1;
-      price += RangeEncoder.getPrice(models[startIndex + m], bit);
+      price += RangeEncoder.getPrice(models[startIndex + m]!, bit)!;
       m = (m << 1) | bit;
     }
     return price;
   }
 
-  static void reverseEncode2(List<int> models, int startIndex,
+  static void reverseEncode2(List<int?> models, int startIndex,
       RangeEncoder rangeEncoder, int numBitLevels, int symbol) {
     var m = 1;
-    for (var i = 0; i < numBitLevels; ++ i) {
+    for (var i = 0; i < numBitLevels; ++i) {
       var bit = symbol & 1;
       rangeEncoder.encode(models, startIndex + m, bit);
       m = (m << 1) | bit;
